+++ Presentation/OpenTgResearcherDesktop/ViewModels/TgUserDetailsViewModel.cs
@@ -5,38 +5,44 @@ public sealed partial class TgUserDetailsViewModel : TgPageViewModelBase
 {
     #region Fields, properties, constructor
 
-	[ObservableProperty]
-	public partial Guid Uid { get; set; } = Guid.Empty!;
-	[ObservableProperty]
-	public partial TgEfUserDto Dto { get; set; } = default!;
+    [ObservableProperty]
+    public partial Guid Uid { get; set; } = Guid.Empty!;
+    [ObservableProperty]
+    public partial TgEfUserDto Dto { get; set; } = default!;
     [ObservableProperty]
     public partial List<long> ListIds { get; set; } = [];
     [ObservableProperty]
-    public partial ObservableCollection<TgEfUserWithMessagesDto> UserWithMessagesDtos { get; set; } = [];
+    public partial ObservableCollection<TgEfChatWithCountDto> ChatsDtos { get; set; } = [];
+    [ObservableProperty]
+    public partial bool IsMessagesLoaded { get; set; }
+    [ObservableProperty]
+    public partial TgEfChatWithCountDto? SelectedChat { get; set; }
 
     public IAsyncRelayCommand LoadDataStorageCommand { get; }
-	public IAsyncRelayCommand ClearViewCommand { get; }
-	public IAsyncRelayCommand UpdateOnlineCommand { get; }
+    public IAsyncRelayCommand ClearViewCommand { get; }
+    public IAsyncRelayCommand UpdateOnlineCommand { get; }
+    public IAsyncRelayCommand<TgEfChatWithCountDto> LoadMessagesCommand { get; }
 
-	public TgUserDetailsViewModel(ITgSettingsService settingsService, INavigationService navigationService, ILoadStateService loadStateService, 
+    public TgUserDetailsViewModel(ITgSettingsService settingsService, INavigationService navigationService, ILoadStateService loadStateService,
         ILogger<TgUserDetailsViewModel> logger) : base(settingsService, navigationService, loadStateService, logger, nameof(TgUserDetailsViewModel))
-	{
-		// Commands
-		ClearViewCommand = new AsyncRelayCommand(ClearViewAsync);
-		LoadDataStorageCommand = new AsyncRelayCommand(LoadDataStorageAsync);
-		UpdateOnlineCommand = new AsyncRelayCommand(UpdateOnlineAsync);
+    {
+        // Commands
+        ClearViewCommand = new AsyncRelayCommand(ClearViewAsync);
+        LoadDataStorageCommand = new AsyncRelayCommand(LoadDataStorageAsync);
+        UpdateOnlineCommand = new AsyncRelayCommand(UpdateOnlineAsync);
+        LoadMessagesCommand = new AsyncRelayCommand<TgEfChatWithCountDto>(LoadMessagesAsync);
     }
 
-	#endregion
+    #endregion
 
-	#region Methods
+    #region Methods
 
-	public override async Task OnNavigatedToAsync(NavigationEventArgs? e) => await LoadStorageDataAsync(async () =>
-		{
-			Uid = e?.Parameter is Guid uid ? uid : Guid.Empty;
-			await LoadDataStorageCoreAsync();
-			await ReloadUiAsync();
-		});
+    public override async Task OnNavigatedToAsync(NavigationEventArgs? e) => await LoadStorageDataAsync(async () =>
+        {
+            Uid = e?.Parameter is Guid uid ? uid : Guid.Empty;
+            await LoadDataStorageCoreAsync();
+            await ReloadUiAsync();
+        });
 
     private async Task ClearViewAsync() => await ContentDialogAsync(ClearDataStorageCore, TgResourceExtensions.AskDataClear(), TgEnumLoadDesktopType.Storage);
 
@@ -44,30 +50,36 @@ public sealed partial class TgUserDetailsViewModel : TgPageViewModelBase
 
     private async Task LoadDataStorageAsync() => await ContentDialogAsync(LoadDataStorageCoreAsync, TgResourceExtensions.AskLoading(), TgEnumLoadDesktopType.Storage);
 
-	private async Task LoadDataStorageCoreAsync()
-	{
-		if (!SettingsService.IsExistsAppStorage) return;
+    // TODO: Add new table CHATS_USERS
+    private async Task LoadDataStorageCoreAsync()
+    {
+        if (!SettingsService.IsExistsAppStorage) return;
 
-        UserWithMessagesDtos.Clear();
+        ChatsDtos.Clear();
         ListIds.Clear();
 
-		Dto = await App.BusinessLogicManager.StorageManager.UserRepository.GetDtoAsync(x => x.Uid == Uid);
-        ListIds = [.. (await App.BusinessLogicManager.StorageManager.MessageRepository.GetListDtosAsync(0, 0, x => x.UserId == Dto.Id)).Select(x => x.SourceId).Distinct()];
+        Dto = await App.BusinessLogicManager.StorageManager.UserRepository.GetDtoAsync(x => x.Uid == Uid);
+        if (Dto is null) return;
 
-        if (ListIds is not null && ListIds.Count != 0)
+        //ListIds = await App.BusinessLogicManager.StorageManager.MessageRepository.GetListDtosAsync(0, 0, x => x.UserId == Dto.Id)).Select(x => x.SourceId).Distinct();
+        ListIds = await App.BusinessLogicManager.StorageManager.MessageRepository
+            .GetListDtosAsync(0, 0, x => x.UserId == Dto.Id)
+            .ContinueWith(t => t.Result.Select(m => m.SourceId).Distinct().ToList());
+        if (ListIds.Count == 0) return;
+
+        foreach (var chatId in ListIds)
         {
-            foreach (var chatId in ListIds)
-            {
-                var messageDtos = await App.BusinessLogicManager.StorageManager.MessageRepository
-                    .GetListDtosAsync(0, 0, x => x.SourceId == chatId && x.UserId == Dto.Id);
-                messageDtos = [.. messageDtos.OrderBy(x => x.Id)];
-                var chatDto = await App.BusinessLogicManager.StorageManager.SourceRepository.GetDtoAsync(x => x.Id == chatId);
-                UserWithMessagesDtos.Add(new TgEfUserWithMessagesDto(Dto, chatDto, messageDtos));
-            }
+            var chatDto = await App.BusinessLogicManager.StorageManager.SourceRepository.GetDtoAsync(x => x.Id == chatId);
+            if (chatDto is null) continue;
+            var countMessages = await App.BusinessLogicManager.StorageManager.MessageRepository.GetCountAsync(x => x.UserId == Dto.Id && x.SourceId == chatDto.Id);
+            ChatsDtos.Add(new TgEfChatWithCountDto(Dto, chatDto, countMessages));
         }
-	}
 
-	private async Task UpdateOnlineAsync() => await ContentDialogAsync(UpdateOnlineCoreAsync, TgResourceExtensions.AskUpdateOnline(), TgEnumLoadDesktopType.Online);
+        // Order
+        ChatsDtos = [.. ChatsDtos.OrderBy(x => x.ChatDto.UserName).ThenBy(x => x.ChatDto.Title)];
+    }
+
+    private async Task UpdateOnlineAsync() => await ContentDialogAsync(UpdateOnlineCoreAsync, TgResourceExtensions.AskUpdateOnline(), TgEnumLoadDesktopType.Online);
 
     private async Task UpdateOnlineCoreAsync() => await LoadOnlineDataAsync(async () =>
     {
@@ -85,5 +97,25 @@ public sealed partial class TgUserDetailsViewModel : TgPageViewModelBase
         }
     });
 
-	#endregion
+    private async Task LoadMessagesAsync(TgEfChatWithCountDto? chat)
+    {
+        if (chat is null) return;
+
+        //var parent = ChatsDtos.FirstOrDefault(u => ChatsDtos.Contains(chat));
+        //if (parent == null || parent.IsMessagesLoaded) return;
+
+        //parent.SelectedChat = chat;
+
+        //var msgs = await App.BusinessLogicManager.StorageManager.MessageRepository
+        //    .GetListDtosAsync(0, 0, x => x.SourceId == chat.Id && x.UserId == parent.UserDto.Id);
+
+        //parent.MessageDtos.Clear();
+        //foreach (var m in msgs.OrderBy(m => m.Id))
+        //    parent.MessageDtos.Add(m);
+
+        //parent.IsMessagesLoaded = true;
+        await Task.CompletedTask;
+    }
+
+    #endregion
 }